Add latest CalSysTakeNarrowbandData.
