+++ python/lsst/ts/standardscripts/auxtel/__init__.py
@@ -1,0 +1,1 @@
+from .calsys_take_narrowband_data import CalSysTakeNarrowbandData
